implement a refresh handler and use it to replace discover tests
When a refresh handler is present, a refresh button will appear in the
test UI pannel and will be called when the button is clicked. The
handler returns a promise that resolves when tests are refreshed,
informing the UI to update progress animations.

The refresh handler is provided a cancellation token which provides the
user with the ability to cancel the operation. This is the reason to
keep the call to ctest as an async spawn instead of syncronous. When
cancellation is requested, the asyncronous process can be signaled to
stop. I tested this by replacing the call to `ctest` with
`sleep infinity` which I was able to cancel in the UI and get a nice
error message in the output window.