Fix the pre-package script
I changed the script name to render_logo.sh, because that is all it
does. I added it to package.json as a prepackage script. I took this
opportunity to make the GH workflow file a bit easier to read, too.
Turns out the package test was broken, too. The files in the
archive are different after using webpack.

closes #27